Tests and implementation for history and history all refactored
The history endpoint now returns a Metadata json object. The requests Client is now passed to the request functions so that the Client.Transport can be overridden to replay previously recorded HTTP responses. Replaying and recording responses removes the need for interface mocking or HTTP server creation. It does add additional complexity when creating these records - it must have valid API secrets which then need to be revoked before committing to github. These must be addressed in documentation, and/or done automatically in github actions by calling itself!