Refactor for tests (#2)
* Create a GenericPrinter instead one per type

* Use requests.Replay for testing the responses from Swyftx

* Remove .idea from repo

* Fix return statement for Generic Printers

`--pretty` was not printing with indent due to missing return statement and it falling through to the standard printer

* Add a metadata type and function to display portfolio data in a cleaner way

Portfolio responses will be made to inform the user of how many records are in total, which page they are on and how many pages remain.

* Extract token getter into its own function

The function has a testing override and error handling included. We need a way to provide a token which matches the replay files as the replay files are a hash of their contents. This means a token which does not match the stored replay files will fail.

* Tests and implementation for history and history all refactored

The history endpoint now returns a Metadata json object. The requests Client is now passed to the request functions so that the Client.Transport can be overridden to replay previously recorded HTTP responses. Replaying and recording responses removes the need for interface mocking or HTTP server creation. It does add additional complexity when creating these records - it must have valid API secrets which then need to be revoked before committing to github. These must be addressed in documentation, and/or done automatically in github actions by calling itself!

* First capture of records for replaying.

* Include client in markets detail and update types/vars

* Remove useless comments

* Add tests

* Add tests for types

* Remove redundant sub command

* Update README.md